First 50 fibonacci series

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,18 +8,18 @@
 
 int main(void)
 {
-	long long int first = 1, second = 2, store;
+	long int first = 1, second = 2, store;
 	int count = 1;
 
-	printf("%lld, ", first);
-	printf("%lld, ", second);
+	printf("%ld, ", first);
+	printf("%ld, ", second);
 
 	while (count <= 50)
 	{
 		store = second;
 		second = second + first;
 		first = store;
-		printf("%lld, ", second);
+		printf("%ld, ", second);
 		count++;
 	}
 	return (0);
